requirements.txt
transformers
torch
streamlit
sumy
bert-extractive-summarizer
scikit-learn
nltk
